refactor: split tests helper methods to separate module
